+++ react/features/base/react/components/index.js
@@ -1,2 +1,3 @@
 export * from './Container';
 export * from './Link';
+export * from './Watermarks';
