+++ react/features/welcome/components/WelcomePage.web.js
@@ -3,14 +3,9 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import { AbstractWelcomePage, mapStateToProps } from './AbstractWelcomePage';
+import { Watermarks } from '../../base/react';
 
-/**
- * The CSS style of the element with CSS class <tt>rightwatermark</tt>.
- */
-const RIGHT_WATERMARK_STYLE = {
-    backgroundImage: 'url(images/rightwatermark.png)'
-};
+import { AbstractWelcomePage, mapStateToProps } from './AbstractWelcomePage';
 
 /* eslint-disable require-jsdoc */
 
@@ -32,7 +27,13 @@ class WelcomePage extends AbstractWelcomePage {
     constructor(props) {
         super(props);
 
-        this._initState();
+        this.state = {
+            ...this.state,
+
+            enableWelcomePage: true,
+            generateRoomnames:
+                interfaceConfig.GENERATE_ROOMNAMES_ON_WELCOME_PAGE
+        };
 
         // Bind event handlers so they are only bound once for every instance.
         this._onDisableWelcomeChange = this._onDisableWelcomeChange.bind(this);
@@ -63,15 +64,13 @@ class WelcomePage extends AbstractWelcomePage {
      */
     render() {
         return (
-            <div>
-                <div id = 'welcome_page'>
-                    {
-                        this._renderHeader()
-                    }
-                    {
-                        this._renderMain()
-                    }
-                </div>
+            <div id = 'welcome_page'>
+                {
+                    this._renderHeader()
+                }
+                {
+                    this._renderMain()
+                }
             </div>
         );
     }
@@ -84,30 +83,6 @@ class WelcomePage extends AbstractWelcomePage {
      */
     _getDomain() {
         return `${window.location.protocol}//${window.location.host}/`;
-    }
-
-    /**
-     * Method that initializes state of the component.
-     *
-     * @returns {void}
-     */
-    _initState() {
-        const showBrandWatermark = interfaceConfig.SHOW_BRAND_WATERMARK;
-        const showJitsiWatermark = interfaceConfig.SHOW_JITSI_WATERMARK;
-
-        this.state = {
-            ...this.state,
-            brandWatermarkLink:
-                showBrandWatermark ? interfaceConfig.BRAND_WATERMARK_LINK : '',
-            enableWelcomePage: true,
-            generateRoomnames:
-                interfaceConfig.GENERATE_ROOMNAMES_ON_WELCOME_PAGE,
-            jitsiWatermarkLink:
-                showJitsiWatermark ? interfaceConfig.JITSI_WATERMARK_LINK : '',
-            showBrandWatermark,
-            showJitsiWatermark,
-            showPoweredBy: interfaceConfig.SHOW_POWERED_BY
-        };
     }
 
     /**
@@ -152,28 +127,6 @@ class WelcomePage extends AbstractWelcomePage {
      */
     _onRoomChange(event) {
         super._onRoomChange(event.target.value);
-    }
-
-    /**
-     * Method that returns brand watermark element if it is enabled.
-     *
-     * @private
-     * @returns {ReactElement|null} Watermark element or null.
-     */
-    _renderBrandWatermark() {
-        if (this.state.showBrandWatermark) {
-            return (
-                <a
-                    href = { this.state.brandWatermarkLink }
-                    target = '_new'>
-                    <div
-                        className = 'watermark rightwatermark'
-                        style = { RIGHT_WATERMARK_STYLE } />
-                </a>
-            );
-        }
-
-        return null;
     }
 
     /**
@@ -239,15 +192,8 @@ class WelcomePage extends AbstractWelcomePage {
 
         return (
             <div id = 'welcome_page_header'>
-                {
-                    this._renderJitsiWatermark()
-                }
-                {
-                    this._renderBrandWatermark()
-                }
-                {
-                    this._renderPoweredBy()
-                }
+                <Watermarks />
+
                 <div id = 'enter_room_container'>
                     <div id = 'enter_room_form'>
                         <div className = 'domain-name'>
@@ -297,47 +243,6 @@ class WelcomePage extends AbstractWelcomePage {
                 <div id = 'header_text' />
             </div>
         );
-    }
-
-    /**
-     * Method that returns jitsi watermark element if it is enabled.
-     *
-     * @private
-     * @returns {ReactElement|null} Watermark element or null.
-     */
-    _renderJitsiWatermark() {
-        if (this.state.showJitsiWatermark) {
-            return (
-                <a
-                    href = { this.state.jitsiWatermarkLink }
-                    target = '_new'>
-                    <div className = 'watermark leftwatermark' />
-                </a>
-            );
-        }
-
-        return null;
-    }
-
-    /**
-     * Renders powered by block if it is enabled.
-     *
-     * @private
-     * @returns {ReactElement|null}
-     */
-    _renderPoweredBy() {
-        if (this.state.showPoweredBy) {
-            return (
-                <a
-                    className = 'poweredby'
-                    href = 'http://jitsi.org'
-                    target = '_new'>
-                    <span data-i18n = 'poweredby' /> jitsi.org
-                </a>
-            );
-        }
-
-        return null;
     }
 
     /**
